Updated WoodsterModule!
This is updating the WoodsterModule for Woodster v2.0.1, v2.1.1, & v2.2.0. These versions of Woodster aren't out yet though but will be soon as I am currently working with my texture artist friend to update some of the textures within Woodster, and thus I decided to update this WoodsterModule in advanced for it.