Update Farmer's Delight cabinet tags (#914)
Every Compat's cabinets (for Farmer's Delight) are currently failing to receive the "mineable/axe" block tag in game. This commit intends to bring further consistency with the tags in Farmer's Delight to hopefully solve the issue, along with some minor cleanup. Farmer's Delight does not create a block tag for cabinets, so that has been removed. The redundant addition of cabinets to the parent "cabinets" item tag has also been removed, as they are already added to the "cabinets/wooden" child.